Backend/requirements.txt
flask==2.3.3
gunicorn==21.2.0
flask-cors==4.0.0
requests==2.31.0
supabase==1.0.1
python-dotenv==1.0.0
Flask-JWT-Extended==4.5.2